Update schema.md
update content; removing individual detail blocks, e.g. 
Data that is easy to edit and combine
{: class="title-color"}
Attributed to William of Ockham, the fourteenth-century logician and philosopher, Occam’s Razor is actually a much older rule of thumb. Also known as the law of parsimony.

Faster dataset selection
{: class="title-color"}
It is usually put into English as the maxim “Do not multiply entities beyond necessity.” The idea is straightforward: don’t concoct a complicated, extravagant theory if you’ve got a simpler one.